refactor(13-1): decompose QuickLogModal into focused sub-components
Split the 548-line QuickLogModal.tsx into 8 focused files under
components/quick-log/: types, SetRow, ExerciseCard, AddExercisePicker,
MuscleGroupSelector, ExerciseLogger, QuickLogModal (shell), and index.
Original file re-exports for backward compatibility.
